+++ src/scaffold/lib/config/rollup.config.lib.js
@@ -1,9 +1,9 @@
 // rollup.config.lib.js
 
 export default {
-  input: 'out-tsc/default-lib.js',
+  input: 'out-tsc/es2015/{{projectName}}.js',
   output: {
-    file: 'dist/{{projectName}}/fesm2015/default-lib.js',
+    file: 'dist/{{projectName}}/fesm2015/{{projectName}}.js',
     format: 'es',
     sourcemap: false
   },